+++ cfdm/mixin/fielddomain.py
@@ -2624,3 +2624,38 @@ class FieldDomain:
             version="1.13.1.0",
             removed_at="5.0.0",
         )  # pragma: no cover
+
+    def to_xarray(self, group=True):
+        """Convert the {{class}} to an `xarray` dataset.
+
+        {{cf_xarray description}}
+
+        Note that ``ds = f.to_xarray()`` is identical to ``ds =
+        cfdm.write(f, fmt='XARRAY')``; and multiple {{class_lower}}s
+        may be written to the same `xarray` dataset with `cfdm.write`
+        (e.g. ``ds = cfdm.write([f, g], fmt='XARRAY')``). Also,
+        `cfdm.write` allows a mixture a mixture of fields and domains
+        to be written to the same `xarray` dataset.
+
+        .. versionadded:: (cfdm) NEXTVERSION
+
+        .. seealso:: `cfdm.write`
+
+        :Parameter:
+
+            group: `bool`, optional
+                If False then create a "flat" dataset, i.e. one with
+                only the root group, regardless of any group structure
+                specified by the netCDF interfaces of the
+                {{class_lower}} and its components. If True (the
+                default) then any sub-groups will be created and
+                populated.
+
+        :Returns:
+
+            {{Returns xarray}}
+
+        """
+        from cfdm.read_write import write
+
+        return write(self, fmt="XARRAY", group=group)