+++ Changelog.rst
@@ -3,6 +3,12 @@ Version NEXTVERSION
 
 **2026-??-??**
 
+* New methods to convert to `xarray`: `cf.Field.to_xarray` and
+  `cf.Domain.to_xarray`
+  (https://github.com/NCAS-CMS/cfdm/issues/394)
+* New output format for `cfdm.write` that creates an `xarray` dataset
+  in memory: ``'XARRAY'``
+  (https://github.com/NCAS-CMS/cfdm/issues/394)
 * Improved and restructured output of `cfdm.Field.dataset_compliance` for
   (at present, preliminary) CF compliance checking, also available
   through a new keyword parameter to `cfdm.read`: ``_noncompliance_report``
@@ -12,6 +18,9 @@ Version NEXTVERSION
   (https://github.com/NCAS-CMS/cfdm/issues/406)
 * Fix bug in `cfdm.write` when writing numpy scalars as Zarr
   attributes (https://github.com/NCAS-CMS/cfdm/issues/408)
+* New optional dependency: ``xarray>=2026.2.0``
+
+----
 
 Version 1.13.1.0
 ----------------